pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-facegen"
dynamic = ["version"]
description = "Generate consistent character portraits with Amazon Bedrock image models"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "Masked Fox Productions" }
]
keywords = [
    "ai",
    "image-generation",
    "portraits",
    "aws",
    "bedrock",
    "game-development",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Games/Entertainment",
]

dependencies = [
    "boto3>=1.34.0",
    "Pillow>=10.0.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "types-Pillow",
]

[project.scripts]
ai-facegen = "ai_facegen.cli:main"

[project.urls]
Homepage = "https://github.com/Masked-Fox-Productions/ai-facemaker"
Documentation = "https://github.com/Masked-Fox-Productions/ai-facemaker#readme"
Repository = "https://github.com/Masked-Fox-Productions/ai-facemaker.git"
Issues = "https://github.com/Masked-Fox-Productions/ai-facemaker/issues"

[tool.hatch.version]
path = "src/ai_facegen/_version.py"

[tool.hatch.build.targets.sdist]
include = [
    "/src",
]

[tool.hatch.build.targets.wheel]
packages = ["src/ai_facegen"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=ai_facegen --cov-report=term-missing"
